+++ hooklib/CMakeLists.txt
@@ -21,6 +21,10 @@ add_library( # Sets the name of the library.
              src/main/cpp/trampoline/trampoline.cpp
              src/main/cpp/trampoline/trampoline_manager.cpp
              src/main/cpp/utils/fake_dlfcn.cpp
+             src/main/cpp/utils/utils.cpp
+             src/main/cpp/utils/offset.cpp
+             src/main/cpp/casts/cast_art_method.cpp
+             src/main/cpp/art/art_method.cpp
              src/main/cpp/trampoline/arch/arm32.S
              src/main/cpp/trampoline/arch/arm64.S
             )
