+++ hooklib/src/main/cpp/art/art_method.cpp
@@ -1,0 +1,40 @@
+//
+// Created by swift on 2019/2/3.
+//
+#include <cstdint>
+#include "../includes/art_method.h"
+#include "../includes/cast_art_method.h"
+
+using namespace art::mirror;
+
+void ArtMethod::setStatic() {
+
+};
+
+void ArtMethod::tryDisableInline() {
+
+}
+
+void ArtMethod::disableInterpreterForO() {
+
+}
+
+void ArtMethod::disableCompilable() {
+
+}
+
+bool ArtMethod::isAbstract() {
+
+}
+
+bool ArtMethod::isNative() {
+
+}
+
+void ArtMethod::setAccessFlags(uint32_t flags) {
+
+}
+
+void ArtMethod::setPrivate() {
+
+}
