--- hooklib/src/main/cpp/casts/art/art.cpp
@@ -1,42 +1,0 @@
-//
-// Created by swift on 2019/2/3.
-//
-#include <cstdint>
-#include "art.h"
-#include <../cast_art_method.h>
-
-namespace art::mirror {
-
-    void ArtMethod::setStatic() {
-
-    };
-
-    void ArtMethod::tryDisableInline() {
-
-    }
-
-    void ArtMethod::disableInterpreterForO() {
-
-    }
-
-    void ArtMethod::disableCompilable() {
-
-    }
-
-    bool ArtMethod::isAbstract() {
-
-    }
-
-    bool ArtMethod::isNative() {
-
-    }
-
-    void ArtMethod::setAccessFlags(uint32_t flags) {
-
-    }
-
-    void ArtMethod::setPrivate() {
-
-    }
-
-}
