+++ hooklib/src/main/cpp/includes/arch_base.h
