+++ hooklib/src/main/cpp/trampoline/arch/arm32.S
@@ -1,4 +1,4 @@
-#include "base.h"
+#include "../../includes/arch_base.h"
 
 #if defined(__arm__)
