+++ hooklib/src/main/cpp/trampoline/arch/arm64.S
@@ -1,4 +1,4 @@
-#include "base.h"
+#include "../../includes/arch_base.h"
 
 /**
 *
