+++ hooklib/src/main/cpp/utils/hide_api.h
@@ -16,13 +16,10 @@ bool (*jitCompileMethod)(void*, void*, void*, bool) = nullptr;
 
 void initHideApi(JNIEnv* env, int SDK_VERSION) {
     if (SDK_VERSION >= 24) {
-        void *handle;
         void *jit_lib;
         if (sizeof(void*) == 8) {
-            handle = fake_dlopen("/system/lib64/libart.so", RTLD_NOW);
             jit_lib = fake_dlopen("/system/lib64/libart-compiler.so", RTLD_NOW);
         } else {
-            handle = fake_dlopen("/system/lib/libart.so", RTLD_NOW);
             jit_lib = fake_dlopen("/system/lib/libart-compiler.so", RTLD_NOW);
         }
         jitCompileMethod = (bool (*)(void *, void *, void *, bool)) fake_dlsym(jit_lib, "jit_compile_method");
